+++ src/cli/help.ts
@@ -6,18 +6,18 @@ import { getEntityName } from '../utils/cliUtils';
 export let additionalHelpMessages: Record<string, string> = {};
 
 export const help: Runner = async (args: Args, ui: UIProvider) => {
-    const cmd = await getEntityName(
-        args._,
-        async () => ''
-    );
-    if (!cmd) {
-        ui.write(helpMessages.help);
-        return;
+    const cmd = args._.length >= 2 ? args._[1].toLowerCase() : '';
+
+    const effectiveHelpMessages: Record<string, string> = {
+        ...additionalHelpMessages,
+        ...helpMessages,
+    };
+
+    for (const k in additionalHelpMessages) {
+        effectiveHelpMessages.help += '\n- ' + k;
     }
-    const key = cmd as keyof typeof helpMessages;
-    if (key && helpMessages[key]) {
-        ui.write(helpMessages[key]);
-    } else {
-        ui.write(helpMessages.help);
-    }
+
+    const helpMessage = cmd in effectiveHelpMessages ? effectiveHelpMessages[cmd] : effectiveHelpMessages['help'];
+
+    ui.write(helpMessage);
 };