+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.31.1] - 2025-04-24
+
+### Fixed
+
+- Fixed build directory creation
+
+## [0.31.0] - 2025-04-24
+
+### Added
+
+- Added Fift output for FunC and Tolk build command
+- Added API reference in code
+
+### Changed
+
+- Improved CLI appearance
+
 ## [0.30.0] - 2025-04-08
 
 ### Fixed